Split project into lib and main
